+++ genre.rb
@@ -10,7 +10,7 @@ class Genre
 
   def add_item(item)
     item.add_genre(self)
-    @items.push(item) unless @items.inclsude?(item)
+    @items.push(item) unless @items.include?(item)
   end
 
   def to_s
