+++ genre_handler.rb
