+++ item.rb
@@ -1,21 +1,19 @@
-class Item
-  attr_reader id
-  attr_accessor genre, author, source, label, publish_date
+require 'date'
+require './labels_manager'
 
-  def initialize(publish_date:, archived: false)
-    @id = Random.rand(1..1_000_000)
+class Item
+  attr_reader :id
+  attr_accessor :genre, :source, :label, :publish_date, :archived
+
+  def initialize(id:, publish_date:, archived: false)
+    @id = id || Random.rand(1..1_000_000)
     @publish_date = Date.parse(publish_date)
     @archived = archived
   end
 
   def add_genre(genre)
     @genre = genre
-    genre.items.push(self) unless genre.items.include?(self)
-  end
-
-  def add_author(author)
-    @author = author
-    author.items.push(self) unless author.items.include?(self)
+    genre&.items&.push(self)
   end
 
   def add_source(source)
@@ -23,7 +21,9 @@ class Item
     source.items.push(self) unless source.items.include?(self)
   end
 
-  def add_label(label)
+  def add_label(label = nil)
+    labels_manager = LabelsManager.new
+    label ||= labels_manager.add_label
     @label = label
     label.items.push(self) unless label.items.include?(self)
   end
@@ -32,10 +32,25 @@ class Item
     @archived = can_be_archived?
   end
 
+  def to_s
+    "ID: #{@id} - Publish Date: #{@publish_date} - Genre: #{@genre&.name} \
+    - Source: #{@source&.name} - Label: #{@label&.name} - Archived? #{@archived}"
+  end
+
+  def to_json(_options = {})
+    {
+      'id' => @id,
+      'publish_date' => @publish_date.strftime('%Y-%m-%d'),
+      'genre_id' => @genre&.id,
+      'source_id' => @source&.id,
+      'label_id' => @label&.id,
+      'archived' => @archived
+    }
+  end
+
   private
 
-  def can_be_archived
-    current_date = Date.today
-    current_date.year - publish_date.year > 10
+  def can_be_archived?
+    (Date.today.year - publish_date.year) > 10
   end
 end
